create app file
